services/notification/requirements.txt
google-cloud-pubsub
bigquery